Replace incidents page
