+++ src/components/Pages.tsx
@@ -5,12 +5,11 @@ import { Page } from "./UI/atoms/Page";
 import { Sidebar } from "./UI/molecules/Sidebar";
 import { Gases } from "./UI/organisms/Gases";
 import { Home } from "./UI/organisms/Home";
-import { Incidents } from "./UI/organisms/Incidents";
 import { Login } from "./UI/organisms/Login";
 import { UserAccount } from "./UI/organisms/UserAccount";
 import { API_URL, getCurrentUser } from "../index";
 import { Locations } from "./UI/organisms/Locations";
-import { IncidentsNext } from "./UI/organisms/IncidentsNext";
+import { Incidents } from "./UI/organisms/Incidents";
 
 const useStyles = makeStyles({
   content: {
@@ -71,12 +70,6 @@ export const Pages: React.FC = () => {
                   path="/incidents"
                   title="Blackline Safety | Incidents"
                   component={Incidents}
-                />
-                <Page
-                  exact
-                  path="/incidentsNext"
-                  title="Blackline Safety | Incidents"
-                  component={IncidentsNext}
                 />
                 <Page
                   exact
